quarkus-artemis-python/requirements.txt
python-qpid-proton
python-dotenv